Implemented a generic diff function on provider base class, so storage integration objects can now be updated/replaced
