+++ pulumi_snowflake/fileformat/file_format_provider.py
@@ -1,4 +1,3 @@
-from pulumi.dynamic import DiffResult
 from pulumi_snowflake import SnowflakeConnectionProvider
 from pulumi_snowflake.snowflakeprovider import IdentifierAttribute
 from pulumi_snowflake.snowflakeprovider.schema_scoped_object_provider import SchemaScopedObjectProvider
@@ -16,19 +15,3 @@ class FileFormatProvider(SchemaScopedObjectProvider):
             IdentifierAttribute("type", True),
             # StringAttribute("comment", False)
         ])
-
-    def diff(self, id, olds, news):
-        fields = ["type", "database", "schema"]
-        changed_fields = []
-
-        for field in fields:
-            if olds.get(field) != news.get(field):
-                changed_fields.append(field)
-
-        if (news.get("name") is not None and olds.get("name") != news.get("name")):
-            changed_fields.append("name")
-        
-        return DiffResult(
-            changes=len(changed_fields) > 0,
-            replaces=changed_fields
-        )
